Refactor thunderstore.repository.admin
Split the thunderstore.repository.admin module into multiple files to
make it easier to navigate.

Refs TS-597
