+++ django/thunderstore/repository/admin/__init__.py
@@ -1,0 +1,5 @@
+from .discord_bot import DiscordUserBotPermissionAdmin
+from .namespace import NamespaceAdmin
+from .package import PackageAdmin
+from .package_rating import PackageRatingAdmin
+from .team import TeamAdmin
